Moving files out of browse/domain directory
Change requested in PR: move these files out of the browse/domain directory.
Plus some mypy fixes.
