+++ tests/test_tex2utf.py
@@ -1,7 +1,7 @@
 """Tests for Tex to UTF8 parsing."""
 from unittest import TestCase
 
-from browse.domain.tex2utf import tex2utf
+from browse.services.util.tex2utf import tex2utf
 
 
 class TextTex2Utf(TestCase):
